Add support for providing extra `copts` strings.
This is important for users with toolchains that will produce
significant warnings when building. We could try to guess at what flags
would silence these warnings (the same way the current code does for
MSVC), but on non-Windows platforms there are multiple different
compilers and compiler versions all with different warning flags.

Instead, simply provide a way for the caller to thread their own flags
through. They can in turn build whatever compiler detection logic is
desired.

It would be possible to just turn off warnings in a blanket way by
passing the flag `-w` on all non-Windows builds. On one hand, this is
a bit simpler. However, it seems rather inelegant and doesn't seem very
sustainable. As an example of where I expect this will break down is for
folks using Clang on Windows.

Happy to have other suggestions for how to configure this -- I don't
claim to be a Bazel expert here.

I'd like to send similar patches for `flex` and `bison` rules as well,
but want to get the pattern worked out here.