Relocate implementations of the public API to single files.
When the `rules_m4` project was first started, the recommended structure
for Bazel rulesets was to have most (or all) of the public API exposed
in a single entry-point file (here, `//m4:m4.bzl`).

Nowadays it's considered better to have the public API in multiple
files, such that the `load()` graph for any given public symbol can be
minimized.

The README will continue to reference the unified `//m4:m4.bzl` since
it's a lot easier to read than a bunch of lengthy `load()` calls.

GitOrigin-RevId: 7c4f57638adb18beebbe8929c1d82178c535f949