Minor tidying-up of workspace-relative vs absolute labels.
The use of absolute labels in `load()` calls was a workaround for a
Bazel bug[0] that was fixed in 2018, before the Bazel v1.0 release.

Using either absolute or relative labels as appropriate makes it
easier to understand whether a given load is expected to happen in
the `@rules_m4` repository or in the root workspace.

[0] https://github.com/bazelbuild/bazel/issues/3115

GitOrigin-RevId: 7e5120adf89230c32569c6d9eb281123f1d48b54